add an example tool to access the lan95xx controller
This commit implements a small (quick and dirty) demonstration,
how to access the controller via userspace USB (libusb) to
do some stuff.
In this case access the chip ID or statistical data.
Both can be used to check the quality of the final solder job.
(Or the ordered chip by that matter...)

How ever it is also possible to extend this tool to make use of the
LAN95XX's free GPIO pins.

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>